+++ engine/scene/core/src/mesh.cpp
@@ -23,7 +23,7 @@ namespace scene
 {
 
 mesh::mesh(scene_manager *scene_manager, const std::shared_ptr<gfx::material> &material,
-           const std::vector<data::vertex_data> &vertex_data, const std::vector<std::uint16_t> &index_data,
+           const std::vector<data::vertex_data> &vertex_data, const std::vector<std::uint32_t> &index_data,
            const std::string &name)
     : scene_object(name,
                    material->sampler_has_alpha() ? render_layer::world_geometry_alpha : render_layer::world_geometry,
